requirements.txt
tensorflow==2.4.0
Pillow==8.0.1
numpy==1.19.4
pre-commit==2.8.2
black==20.8b1
isort==5.6.4
pytest==5.4.3
pytest-sugar==0.9.4